Builds and runs, remaining correctness bugs

+++ crates/ida/src/constants.rs
@@ -11,19 +11,19 @@ pub const MXSTEP_DEFAULT: usize = 500;
 //---------------------
 
 /// max number of convergence failures allowed
-pub const MXNCF: u32 = 10;
+pub const MXNCF: usize = 10;
 /// max number of error test failures allowed
-pub const MXNEF: u32 = 10;
+pub const MXNEF: usize = 10;
 /// max. number of h tries in IC calc.
-pub const MAXNH: u32 = 5;
+pub const MAXNH: usize = 5;
 /// max. number of J tries in IC calc.
-pub const MAXNJ: u32 = 4;
+pub const MAXNJ: usize = 4;
 /// max. Newton iterations in IC calc.
-pub const MAXNI: u32 = 10;
+pub const MAXNI: usize = 10;
 /// Newton convergence test constant
 pub const EPCON: f64 = 0.33;
 /// max backtracks per Newton step in IDACalcIC
-pub const MAXBACKS: u32 = 100;
+pub const MAXBACKS: usize = 100;
 /// constant for updating Jacobian/preconditioner
 pub const XRATE: f64 = 0.25;
